Move Kinepolis Mapping Files to subFolder
