+++ src/Kinepolis/Mapping/MappingRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Mapping;
 
 interface MappingRepository
 {
