+++ src/Kinepolis/Mapping/MovieMappingRepository.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Mapping;
 
 use Doctrine\DBAL\Connection;
